pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fpdf>=1.7.2",
    "google-genai>=1.42.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "openai>=2.3.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "sift-stack-py>=0.9.1",
    "snowflake-connector-python>=4.0.0",
    "statsmodels>=0.14.5",
    "streamlit-lottie>=0.0.5",
    "streamlit>=1.50.0",
    "redis>=6.4.0",
]
